moved the interaction with students and marks to a separate evaluation package
