+++ src/main/java/evaluation/dao/MarkDaoImpl.java
@@ -1,7 +1,7 @@
-package dao;
+package evaluation.dao;
 
 import com.sun.istack.NotNull;
-import models.Mark;
+import evaluation.model.Mark;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
